fall back to "Bergelson Lab" for the nav title
On narrow screens/ mobile, the top menu disappears and hides under a hamburger (triple bar icon) and a semi-transparent title appears on its right. The text for that title is taken from the logo (defined as element with "logo" id). But since most of the pages do not have such an element, it would just become "undefined". Now, any page without a logo element, will just use "Bergelson Lab" instead.